:lipstick: style: requestDto class 분리
