--- src/main/java/com/dojinyou/devcourse/voucherapplication/voucher/Entity/Voucher.java
@@ -1,13 +1,0 @@
-package com.dojinyou.devcourse.voucherapplication.voucher.Entity;
-
-
-public abstract class Voucher {
-    public static class Request {
-        private final VoucherType voucherType;
-        private final VoucherAmount voucherAmount;
-        public Request(VoucherType voucherType, VoucherAmount voucherAmount) {
-            this.voucherType = voucherType;
-            this.voucherAmount = voucherAmount;
-        }
-    }
-}
