+++ src/main/java/com/dojinyou/devcourse/voucherapplication/voucher/entity/Voucher.java
@@ -1,0 +1,5 @@
+package com.dojinyou.devcourse.voucherapplication.voucher.entity;
+
+
+public abstract class Voucher {
+}
